contrib: bone-flash-tools: parametrize mount points, add check for LEDs

+++ contrib/bone-flash-tool/emmc.sh
@@ -5,6 +5,9 @@
 
 export PATH=$PATH:/bin:/sbin:/usr/bin:/usr/sbin
 
+PART1MOUNT="/media/1"
+PART2MOUNT="/media/2"
+
 cd /build
 
 if [ -e /eeprom.dump ] ; then
@@ -16,48 +19,48 @@ echo "Paritioning eMMC"
 ./mkcard.sh /dev/mmcblk1
 
 echo "Mounting partitions"
-mkdir -p /media/1
-mkdir -p /media/2
-mount /dev/mmcblk1p1 /media/1 -o async
-mount /dev/mmcblk1p2 /media/2 -o async,noatime
+mkdir -p ${PART1MOUNT}
+mkdir -p ${PART2MOUNT}
+mount /dev/mmcblk1p1 ${PART1MOUNT} -o relatime
+mount /dev/mmcblk1p2 ${PART2MOUNT} -o async,noatime
 
 echo "Copying bootloader files"
-cp MLO u-boot.img /media/1
-echo "mmcdev=1" > /media/1/uEnv.txt
-echo "optargs=quiet" >> /media/1/uEnv.txt
+cp MLO u-boot.img ${PART1MOUNT}
+echo "optargs=quiet" >> ${PART1MOUNT}/uEnv.txt
 
 sync
 
 echo "Extracting rootfs"
-tar zxf Angstrom-Cloud9-IDE-GNOME-eglibc-ipk-v2012.12-beaglebone.rootfs.tar.gz -C /media/2
+tar zxf Angstrom-Cloud9-IDE-GNOME-eglibc-ipk-v2012.12-beaglebone.rootfs.tar.gz -C ${PART2MOUNT}
 
 echo "Populating VFAT partition"
-if [ -d /media/2/usr/share/beaglebone-getting-started ] ; then
-	cp -r /media/2/usr/share/beaglebone-getting-started/* /media/1
+if [ -d ${PART2MOUNT}/usr/share/beaglebone-getting-started ] ; then
+	cp -r ${PART2MOUNT}/usr/share/beaglebone-getting-started/* ${PART1MOUNT}
 fi
 
-umount /media/1
+umount ${PART1MOUNT}
 
 echo "Generating machine ID"
-systemd-nspawn -D /media/2 /bin/systemd-machine-id-setup
+systemd-nspawn -D ${PART2MOUNT} /bin/systemd-machine-id-setup
 
 echo "Running Postinsts"
 cpufreq-set -g performance
-systemd-nspawn -D /media/2 /usr/bin/opkg-cl configure
+systemd-nspawn -D ${PART2MOUNT} /usr/bin/opkg-cl configure
 cpufreq-set -g ondemand
 
 #echo "Setting timezone to Europe/Paris"
-#systemd-nspawn -D /media/2 /usr/bin/timedatectl set-timezone Europe/Paris
+#systemd-nspawn -D ${PART2MOUNT} /usr/bin/timedatectl set-timezone Europe/Paris
 
-rm /media/2/etc/pam.d/gdm-autologin
+rm ${PART2MOUNT}/etc/pam.d/gdm-autologin
 
-umount /media/2
+umount ${PART2MOUNT}
 
 sync
 
-echo default-on > /sys/class/leds/beaglebone\:green\:usr0/trigger
-echo default-on > /sys/class/leds/beaglebone\:green\:usr1/trigger
-echo default-on > /sys/class/leds/beaglebone\:green\:usr2/trigger
-echo default-on > /sys/class/leds/beaglebone\:green\:usr3/trigger
-
+if [ -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
+	echo default-on > /sys/class/leds/beaglebone\:green\:usr0/trigger
+	echo default-on > /sys/class/leds/beaglebone\:green\:usr1/trigger
+	echo default-on > /sys/class/leds/beaglebone\:green\:usr2/trigger
+	echo default-on > /sys/class/leds/beaglebone\:green\:usr3/trigger
+fi
